Refactor CLI to use Core
